🩹Chore : 기능별 파일 분리 및 주석 제거
